+++ build-logic/src/main/kotlin/com/gabrielfeo/integration-test-suite.gradle.kts
@@ -13,26 +13,14 @@ testing {
     }
 }
 
-tasks.named("check") {
-    dependsOn("integrationTest")
-}
-
 kotlin {
     target {
         val main by compilations.getting
         val integrationTest by compilations.getting
-        val test by compilations.getting
-        val testFixtures by compilations.getting
-        test.associateWith(main)
-        test.associateWith(testFixtures)
         integrationTest.associateWith(main)
-        integrationTest.associateWith(testFixtures)
-        testFixtures.associateWith(main)
     }
 }
 
-// TODO Unapply test-fixtures and delete the source set, since we're not publishing it?
-components.named<AdhocComponentWithVariants>("java") {
-    withVariantsFromConfiguration(configurations["testFixturesApiElements"]) { skip() }
-    withVariantsFromConfiguration(configurations["testFixturesRuntimeElements"]) { skip() }
+tasks.named("check") {
+    dependsOn("integrationTest")
 }
