+++ library/src/testFixtures/kotlin/com/gabrielfeo/develocity/api/Resources.kt
